start to add some structure
