+++ src/api/types.ts
@@ -1,0 +1,6 @@
+import { ISupportCodeLibrary } from '../support_code_library_builder/types'
+
+export interface IRunResult {
+  success: boolean
+  support: ISupportCodeLibrary
+}
