+++ src/configuration/discoverConfiguration.ts
@@ -1,0 +1,5 @@
+import { IRunConfiguration } from './types'
+
+export async function discoverConfiguration(): Promise<IRunConfiguration> {
+  return undefined
+}
